Refact EditProfileForm component
